Move dashboard components in a separate folder
